sorting Algorithms

+++ 3-quick_sort.c
@@ -65,5 +65,5 @@ void quick_sort(int *array, size_t size)
 {
 	if (array == NULL || size < 2)
 		return;
-	quick_sort_recursion(array, size, 0, (int)size - 1);
+	quick_sort_recursion(array, size, 0, size - 1);
 }
